+++ lib/src/widgets/toolbar/buttons/hearder_style/select_header_style_dropdown_button.dart
@@ -90,14 +90,14 @@ class _QuillToolbarSelectHeaderStyleDropdownButtonState
   }
 
   double get iconSize {
-    final baseFontSize = context.quillToolbarBaseButtonOptions?.globalIconSize;
+    final baseFontSize = context.quillToolbarBaseButtonOptions?.iconSize;
     final iconSize = widget.options.iconSize;
     return iconSize ?? baseFontSize ?? kDefaultIconSize;
   }
 
   double get iconButtonFactor {
     final baseIconFactor =
-        context.quillToolbarBaseButtonOptions?.globalIconButtonFactor;
+        context.quillToolbarBaseButtonOptions?.iconButtonFactor;
     final iconButtonFactor = widget.options.iconButtonFactor;
     return iconButtonFactor ?? baseIconFactor ?? kDefaultIconButtonFactor;
   }