+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
-## 9.1.2
+## 9.2.0
 * Require minimum version `6.0.0` of `flutter_keyboard_visibility` to fix some build issues with Android Gradle Plugin 8.2.0
 * Add on image clicked in `flutter_quill_extensions` callback
+* Deprecate `globalIconSize` and `globalIconButtonFactor`, use `iconSize` and `iconButtonFactor` instead
 
 ## 9.1.1
 * Require `super_clipboard` minimum version `0.8.1` to fix some bug with Linux build failure